Revert "LPS-154708 Move BundleStartStopLogger to portal-bootstrap"
This reverts commit 8bec7cfe38f329f4cbb0510ecf248958fc0e2f6b.
