+++ modules/core/portal-equinox-log-bridge/src/main/java/com/liferay/portal/equinox/log/bridge/internal/BundleStartStopLogger.java
@@ -12,55 +12,24 @@
  * details.
  */
 
-package com.liferay.portal.bootstrap.log;
-
-import com.liferay.petra.string.StringBundler;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
+package com.liferay.portal.equinox.log.bridge.internal;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.osgi.framework.Bundle;
-import org.osgi.framework.BundleContext;
 import org.osgi.framework.BundleEvent;
-import org.osgi.framework.InvalidSyntaxException;
-import org.osgi.framework.ServiceReference;
 import org.osgi.framework.SynchronousBundleListener;
-import org.osgi.util.tracker.ServiceTracker;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * @author Raymond Augé
  */
 public class BundleStartStopLogger implements SynchronousBundleListener {
 
-	public BundleStartStopLogger(BundleContext bundleContext)
-		throws InvalidSyntaxException {
-
-		ServiceTracker<Object, Void> serviceTracker =
-			new ServiceTracker<Object, Void>(
-				bundleContext,
-				bundleContext.createFilter(
-					StringBundler.concat(
-						"(&(objectClass=",
-						"com.liferay.portal.kernel.module.framework.",
-						"ModuleServiceLifecycle)",
-						"(module.service.lifecycle=portal.initialized))")),
-				null) {
-
-				@Override
-				public Void addingService(
-					ServiceReference<Object> serviceReference) {
-
-					_portalStarted.set(true);
-
-					close();
-
-					return null;
-				}
-
-			};
-
-		serviceTracker.open();
+	public BundleStartStopLogger(AtomicBoolean portalStarted) {
+		_portalStarted = portalStarted;
 	}
 
 	@Override
@@ -68,32 +37,32 @@ public class BundleStartStopLogger implements SynchronousBundleListener {
 		Bundle bundle = bundleEvent.getBundle();
 
 		if (bundle.getSymbolicName() == null) {
-			_log.error(bundle.getLocation() + " has a null symbolic name");
+			_log.error("{} has a null symbolic name", bundle.getLocation());
 		}
 
 		if (_portalStarted.get()) {
 			if (_log.isInfoEnabled()) {
 				if (bundleEvent.getType() == BundleEvent.STARTED) {
-					_log.info("STARTED " + bundle);
+					_log.info("STARTED {}", bundle);
 				}
 				else if (bundleEvent.getType() == BundleEvent.STOPPED) {
-					_log.info("STOPPED " + bundle);
+					_log.info("STOPPED {}", bundle);
 				}
 			}
 		}
 		else if (_log.isDebugEnabled()) {
 			if (bundleEvent.getType() == BundleEvent.STARTED) {
-				_log.debug("STARTED " + bundle);
+				_log.debug("STARTED {}", bundle);
 			}
 			else if (bundleEvent.getType() == BundleEvent.STOPPED) {
-				_log.debug("STOPPED " + bundle);
+				_log.debug("STOPPED {}", bundle);
 			}
 		}
 	}
 
-	private static final Log _log = LogFactoryUtil.getLog(
+	private static final Logger _log = LoggerFactory.getLogger(
 		BundleStartStopLogger.class);
 
-	private final AtomicBoolean _portalStarted = new AtomicBoolean();
+	private final AtomicBoolean _portalStarted;
 
 }
